quantv3/requirements.txt
pandas>=2.0.0
numpy>=1.24.0
psycopg2-binary>=2.9.0
TA-Lib>=0.4.28
backtrader>=1.9.78
deap>=1.4.1
matplotlib>=3.7.0
scikit-learn>=1.3.0
